+++ webui/Cargo.toml
@@ -8,6 +8,10 @@ edition = "2021"
 path = "../graphql"
 version = "0.1.0"
 
+[dependencies.music-player-settings]
+path = "../settings"
+version = "0.1.0"
+
 [dependencies]
 actix-web = "4.2.1"
 async-graphql = "4.0.15"
@@ -15,3 +19,5 @@ async-graphql-actix-web = "4.0.15"
 mime_guess = "2.0.4"
 owo-colors = "3.5.0"
 rust-embed = { version = "6.4.1", features = ["actix"] }
+actix-files = "0.6.2"
+dirs = "4.0.0"
